[Enhancement] Decouple export map components from exportMap element
